# 25mongodb/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/deleteallmovie": {
            "delete": {
                "description": "Delete all movies from the watchlist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Delete all movies",
                "responses": {
                    "200": {
                        "description": "Number of deleted movies",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "/api/movie": {
            "post": {
                "description": "Add a new movie to the watchlist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Create a movie",
                "parameters": [
                    {
                        "description": "Movie to create",
                        "name": "movie",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Netflix"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Netflix"
                        }
                    }
                }
            }
        },
        "/api/movie/{id}": {
            "put": {
                "description": "Update a movie's watched status to true",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Mark a movie as watched",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Movie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movie ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a movie from the watchlist by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Delete a movie",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Movie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movie ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/movies": {
            "get": {
                "description": "Get all movies from the watchlist",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Get all movies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Netflix"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Netflix": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "movie": {
                    "type": "string"
                },
                "watched": {
                    "type": "boolean"
                }
            }
        }
    }
}